add code expand button
